update pie charts to import pie options from chart options file
