+++ src/components/Containers.js
@@ -1,28 +1,10 @@
 import React from 'react';
-import Chartist from 'chartist';
 import ChartistGraph from 'react-chartist';
 import chartOptions from '../helpers/ChartOptions';
 
 
 class Containers extends React.Component {
   render() {
-    const pieOptions = {
-      donut: true,
-      donutWidth: 50,
-      startAngle: 270,
-      total: 200,
-      showLabel: false,
-      height: '300px',
-      plugins: [
-        Chartist.plugins.tooltip({
-          appendToBody: false,
-          transformTooltipTextFnc: function(value) {
-            return value + '%';
-          }
-        }),
-      ]
-    };
-
 
     const pieChartData = {
       labels: [
@@ -41,7 +23,7 @@ class Containers extends React.Component {
         <ChartistGraph
           className={'ct-gauge'}
           data={pieChartData}
-          options={pieOptions}
+          options={chartOptions.gaugeOptions}
           responsiveOptions={chartOptions.gaugeResponsiveOptions}
           type={'Pie'} />
         <span className='gauge-label'>58%</span><p className="montserrat-lc small"> of responses indicated they are using container tools like Docker and Kubernetes.</p>
