+++ src/components/RailsUpdates.js
@@ -1,29 +1,9 @@
 import React from 'react';
-import Chartist from 'chartist';
 import ChartistGraph from 'react-chartist';
-import ChartistPluginLegend from 'chartist-plugin-legend';
 import chartOptions from '../helpers/ChartOptions';
-
 
 class RailsUpdates extends React.Component {
   render() {
-    const pieOptions = {
-      donut: true,
-      showLabel: false,
-      donutWidth: 80,
-      height: '300px',
-      plugins: [
-        ChartistPluginLegend({
-          clickable: false
-        }),
-        Chartist.plugins.tooltip({
-          appendToBody: false,
-          transformTooltipTextFnc: function(value) {
-            return value + '%';
-          }
-        }),
-      ]
-    };
 
     const releasesPieChartData = {
       labels: [
@@ -40,12 +20,11 @@ class RailsUpdates extends React.Component {
       ]
     };
 
-
     return (
       <div>
         <ChartistGraph
           data={releasesPieChartData}
-          options={pieOptions}
+          options={chartOptions.pieOptions}
           responsiveOptions={chartOptions.pieResponsiveOptions}
           type={'Pie'} />
       </div>
